Create the drive_home_page_view
